QBuilder/requirements.txt
git+git@github.com:black-puppydog/python-bitcoinrpc.git@p34-compatablity#egg=python-bitcoinrpc
cassandra-driver

